set update logic to highest level (MapReducer) and implement Ignite-Backend
